+++ src/main/java/com/ga/gaent/service/EdocService.java
@@ -43,87 +43,6 @@ public class EdocService {
         return edocMapper.selectApprover();
     }
     
-    /*
-     * @author : 정건희, 조인환
-     * @since : 2024. 07. 15.
-     * Description : 전자결재(전자결재 문서, 문서양식, 결재선) 데이터 입력
-     */
-    public int insertEdoc(
-            EdocRequestDTO edocRequestDTO,
-            EdocFormTypeDTO edocFormTypeDTO) {
-        
-        // 공통으로 들어가는 전자결재 문서
-        int edocResult = edocMapper.insertEdoc(edocRequestDTO);
-        
-        Integer edocNum = edocRequestDTO.getEdocNum();
-        edocRequestDTO.setEdocNum(edocNum);
-        edocFormTypeDTO.setEdocNum(edocNum);
-        
-        
-        String[] approvers = edocRequestDTO.getApprover();
-        String[] apprOrders = edocRequestDTO.getApprOrder();
-        
-        int approvalResult = -1;
-        
-        for(int i = 0; i < approvers.length; i++) {
-            Map<String, Object> edocMap = new HashMap<>();
-            edocMap.put("edocNum", edocNum);
-            edocMap.put("approver", approvers[i]);
-            edocMap.put("apprOrder", apprOrders[i]);
-            // 공통으로 들어가는 결재선
-            approvalResult = edocMapper.insertApprover(edocMap);
-        }
-        
-        String edocType = edocFormTypeDTO.getEdocType();
-        
-        int insertEdocForm = -1;
-        if(edocType.equals("0")){
-            System.out.println("기안서 제출");
-            insertEdocForm = edocMapper.insertEdocDraft(edocFormTypeDTO);
-        }else if(edocType.equals("1")){
-            System.out.println("휴가신청서 제출");
-            insertEdocForm = edocMapper.insertEdocVacation(edocFormTypeDTO);
-        }else if(edocType.equals("2")){
-            System.out.println("지출결의서 제출");
-            insertEdocForm = edocMapper.insertEdocProject(edocFormTypeDTO);
-        }else if(edocType.equals("3")){
-            System.out.println("경조사지출결의서 제출");
-            insertEdocForm = edocMapper.insertEdocEvent(edocFormTypeDTO);
-        }else if(edocType.equals("4")){
-            System.out.println("보고서 제출");
-            insertEdocForm = edocMapper.insertEdocReport(edocFormTypeDTO);
-        }
-        
-        
-        if(edocResult != 1 || insertEdocForm != 1) {
-            throw new RuntimeException("전자결재 입력을 실패했습니다.");
-        }
-        
-        return 1;
-    }
-    
-    /*
-     * @author : 정건희
-     * @since : 2024. 07. 15.
-     * Description : 전자결재 첨부파일 데이터 입력
-     */
-    public int insertEdocFile(EdocRequestDTO edocRequestDTO) {
-        
-        String originalFileName = edocRequestDTO.getFileName().getOriginalFilename();
-        String fileName = (UUID.randomUUID().toString()).replace("-", "");
-        String fileType = edocRequestDTO.getFileName().getContentType();
-        /* String file = fileName + fileType; */
-        double fileSize = edocRequestDTO.getFileName().getSize();
-        
-        Map<String, Object> insertFile = new HashMap<>();
-        
-        insertFile.put("originalFileName", originalFileName);
-        insertFile.put("fileName", fileName);
-        insertFile.put("fileType", fileType);
-        insertFile.put("fileSize", fileSize);
-        
-        return edocMapper.insertEdocFile(insertFile);
-    }
     
     /*
      * @author : 정건희
@@ -226,10 +145,10 @@ public class EdocService {
 //        String code1 =  (String)resultMap.get("approver1");
 //        String code2 =  (String)resultMap.get("approver2");
         
-        
-        resultMap.put("approver1",  edocMapper.findKorName((String)resultMap.get("approver1")));
-        resultMap.put("approver2",  edocMapper.findKorName((String)resultMap.get("approver2")));
-        
+        if(resultMap != null) {
+        resultMap.put("approverName1",  edocMapper.findKorName((String)resultMap.get("approver1")));
+        resultMap.put("approverName2",  edocMapper.findKorName((String)resultMap.get("approver2")));
+        }
         
         
         return resultMap;
@@ -260,33 +179,6 @@ public class EdocService {
     
     
     
-    
-    
-    
-    
-    //  결재,반려처리
-    public int updateEdocProcess(int empCode,String edocNum,String apprReason, int request) {
-        
-        String requestEnum =  String.valueOf(request);
-        Map<String,Object>map = new HashMap<>();
-        map.put("empCode", empCode);
-        map.put("edocNum", edocNum);
-        map.put("apprReason", apprReason);
-        map.put("request", requestEnum);
-        
-        if(request < 0) {
-            map.put("apprStatus", "-1" );
-        }else if(request > 0){
-            map.put("apprStatus", "1" );
-        }
-        
-        int s1 = edocMapper.updateEdocApprovalStatus(map); // 결재선테이블
-        log.debug(TeamColor.YELLOW + "결재선" + s1 + TeamColor.RESET);
-        
-        int s2 = edocMapper.updateEdocStatus(map); // 전자문서공통테이블
-        log.debug(TeamColor.YELLOW + "공통" + s2 + TeamColor.RESET);
-        return 1;
-    }
     
     // 개인이 제출한 대기,진행,반려문서
     public List<EdocVO>selectMyEdocSubmitList(String empCode, int request){
